Fix static analysis const reference performance warnings.
Credit to cmangos https://github.com/cmangos/mangos-classic/commit/14f57745471f0c3a3b262e168b2875f1a640daa3

+++ src/game/Conditions.cpp
@@ -544,11 +544,11 @@ bool inline ConditionEntry::Evaluate(WorldObject const* target, Map const* map, 
             switch (m_value2)
             {
                 case 0:
-                    return visualRank == m_value1;
+                    return visualRank == int8(m_value1);
                 case 1:
-                    return visualRank >= m_value1;
+                    return visualRank >= int8(m_value1);
                 case 2:
-                    return visualRank <= m_value1;
+                    return visualRank <= int8(m_value1);
             }
             return false;
         }
